moved domain classes to empleodotnet.core adjusted namespaces for domain classes adjusted .net framework to library projects to match main project(mvc)
