+++ EmpleoDotNet.Core/Domain/JobOpportunity.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace EmpleoDotNet.Models
+namespace EmpleoDotNet.Core.Domain
 {
     public class JobOpportunity : EntityBase
     {
