+++ EmpleoDotNet.sln
@@ -1,11 +1,17 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.24720.0
+VisualStudioVersion = 14.0.23107.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmpleoDotNet", "EmpleoDotNet\EmpleoDotNet.csproj", "{AFE7FC0B-2D9D-4260-A25F-6B523F9A6A0F}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmpleoDotNet.Tests", "EmpleoDotNet.Tests\EmpleoDotNet.Tests.csproj", "{F76D7DB5-E032-42C3-BDEF-84A1405B6633}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmpleoDotNet.Core", "EmpleoDotNet.Core\EmpleoDotNet.Core.csproj", "{8677B534-ED45-4526-9A6A-FC08F35BFE8C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmpleoDotNet.Data", "EmpleoDotNet.Data\EmpleoDotNet.Data.csproj", "{32141584-D6C9-44C2-AC12-7F21F6043EE5}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmpleoDotNet.Repository", "EmpleoDotNet.Repository\EmpleoDotNet.Repository.csproj", "{4E677DEC-45CF-4E88-9D49-2C0BE16FFACE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +27,18 @@ Global
 		{F76D7DB5-E032-42C3-BDEF-84A1405B6633}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F76D7DB5-E032-42C3-BDEF-84A1405B6633}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F76D7DB5-E032-42C3-BDEF-84A1405B6633}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8677B534-ED45-4526-9A6A-FC08F35BFE8C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8677B534-ED45-4526-9A6A-FC08F35BFE8C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8677B534-ED45-4526-9A6A-FC08F35BFE8C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8677B534-ED45-4526-9A6A-FC08F35BFE8C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{32141584-D6C9-44C2-AC12-7F21F6043EE5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{32141584-D6C9-44C2-AC12-7F21F6043EE5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{32141584-D6C9-44C2-AC12-7F21F6043EE5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{32141584-D6C9-44C2-AC12-7F21F6043EE5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4E677DEC-45CF-4E88-9D49-2C0BE16FFACE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4E677DEC-45CF-4E88-9D49-2C0BE16FFACE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4E677DEC-45CF-4E88-9D49-2C0BE16FFACE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4E677DEC-45CF-4E88-9D49-2C0BE16FFACE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
